smart-framework/src/main/java/org/smart4j/framework/util/PropsUtil.java
package org.smart4j.framework.util;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.io.IOException;
import java.io.InputStream;
import java.util.Properties;

/**
 * Properties 工具类
 */
public final class PropsUtil {

    private static final Logger LOGGER = LoggerFactory.getLogger(PropsUtil.class);

    public static Properties loadProps(String fileName) {
        Properties props;
        try (InputStream is = Thread.currentThread().getContextClassLoader().getResourceAsStream(fileName)) {
            if (is == null) {
                LOGGER.error("fileName: " + fileName);
                throw new RuntimeException("File don't exist");
            }
            props = new Properties();
            props.load(is);
        } catch (IOException e) {
            LOGGER.error("fileName: " + fileName);
            throw new RuntimeException("Load properties file failure", e);
        }
        return props;
    }

    public static String getString(Properties props, String key) {
        return getString(props, key, "");
    }

    public static String getString(Properties props, String key, String defaultValue) {
        String value = defaultValue;
        if (props.containsKey(key)) {
            value = props.getProperty(key);
        }
        return value;
    }

}
